+++ onnx2mlx/__init__.py
@@ -3,5 +3,6 @@
 #   Copyright © 2026 NatML Inc. All Rights Reserved.
 #
 
-from .converter import onnx2mlx, ConvertContext
+from .convert import convert_to_mlx, ConvertContext
+from .export import export_to_mlx
 from .version import __version__